requirement.txt
fastapi
uvicorn
azure-storage-blob
python-dotenv
click
requests
pytest